rag/requirements.txt
faiss-cpu==1.8.0
sentence-transformers==2.7.0
uvicorn==0.30.1
fastapi==0.111.0
PyPDF2==3.0.1


